# [SPARK-21217][SQL] Support ColumnVector.Array.to<type>Array()