@cloud-fan 2017-06-28
Hi @kiszk , thanks for working on it, do you have any performance numbers?