@kiszk 2017-06-28
I have not had performance numbers. I will do performance experiments.